Update and rename test to Randomise Outbound CLI - Global
Can be used to add a pool of numbers for different Countries, and when a call is made, a random number will be chosen to be displayed as the Caller ID. This number will then be stored in a Agent_Number__c field on the Call Reporting Record (will need to be added first)

If the at least 1 number exists for the called country, this will be set as the CLI and will return true. If no number is found (country is not in the table), then the script will return false.